pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "discord-zulip-bridge"
version = "0.1.0"
description = "Bridge messages between one Discord channel and one Zulip topic"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "discord.py>=2.4.0",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.1",
]

[project.scripts]
discord-zulip-bridge = "discord_zulip_bridge.bridge_app:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["discord_zulip_bridge*"]
